requirements.txt
flask
flask_sqlalchemy
passlib
datetime
pyjwt==2.8.0
werkzeug
uuid
gunicorn
# gevent
pymysql==1.1.0